+++ pane.go
@@ -174,21 +174,18 @@ func (pane *Pane) Resize(ws *pty.Winsize) {
 }
 
 // pane.Restore sends the panes' visible lines line to a data channel
+// the data channel is specified by it index in the pane.dcs slice
 // the function does nothing if it's given a nil size or the current size
-func (pane *Pane) Restore(d *webrtc.DataChannel) {
+func (pane *Pane) Restore(dIdx int) {
 	if pane.st != nil {
 		Logger.Infof("Sending scrren dump to %d", pane.Id)
-		b, l := STDump(pane.st)
-		if l > 0 {
-			d.Send(b)
-			// position the cursor
-			ps := fmt.Sprintf("\x1b[%d;%dH",
-				int(pane.st.c.y)+1, int(pane.st.c.x)+1)
-			d.Send([]byte(ps))
-		} else {
-			Logger.Info("not restoring as dump len is 0")
-		}
+		dc := STDumpContext{pane.Id, dIdx}
+		STDump(pane.st, &dc)
+		// position the cursor
+		ps := fmt.Sprintf("\x1b[%d;%dH",
+			int(pane.st.c.y)+1, int(pane.st.c.x)+1)
+		pane.dcs[dIdx].Send([]byte(ps))
 	} else {
-		Logger.Info("not restoring as st is null")
+		Logger.Warn("not restoring as st is null")
 	}
 }