+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixes
+
+- Crash on large pane restore
+
 ### Adds
+
 - Screen buffer & cursor position restore. Monchrome & effectless for now
 
 ## [0.3.0] - 2020-10-04