+++ CHANGELOG.md
@@ -5,16 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.4.0] - 2020-10-14
 
 ### Fixes
 
-- Crash on large pane restore
 - Linter based code beutification
 
 ### Adds
 
-- Screen buffer & cursor position restore. Monchrome & effectless for now
+- Screen buffer & cursor position restore. Monchrome & plain for now
 
 ## [0.3.0] - 2020-10-04